Added Carlos Blake2s script. Increased supported integer size. Reduced cloning when doing function inlining. Co-authored-by: carlo modica <carlo.modica@protonmail.com>

+++ src/ast.rs
@@ -164,7 +164,7 @@ pub enum Pattern {
     As(Box<Pattern>, Variable),
     Product(Box<Pattern>, Box<Pattern>),
     Variable(Variable),
-    Constant(i32),
+    Constant(i128),
 }
 
 impl Pattern {
@@ -282,7 +282,7 @@ pub enum Expr {
     Infix(InfixOp, Box<TExpr>, Box<TExpr>),
     Negate(Box<TExpr>),
     Application(Box<TExpr>, Box<TExpr>),
-    Constant(i32),
+    Constant(i128),
     Variable(Variable),
     Function(Function),
     Intrinsic(Intrinsic),
